ai-training/requirements.txt
# Web
fastapi==0.92.0
starlette==0.25.0
uvicorn[standard]==0.22.0
python-multipart==0.0.9

# Core typing pin to satisfy TensorFlow
typing-extensions==4.5.0

# ML / DL (CPU)
# Note: On Windows, TensorFlow 2.13.x requires numpy<=1.24.3 and >=1.22
tensorflow==2.13.1
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.10.1
pillow==10.2.0
opencv-python==4.9.0.80

# AWS / S3
boto3==1.34.34
botocore==1.34.34

# Supabase + HTTP
supabase>=2.6.0
# httpx >=0.26 so Client accepts `proxy` kwarg used by gotrue/supabase
httpx>=0.27,<0.28
requests==2.31.0
python-dotenv==1.0.1

# Serialization
orjson==3.9.15

# Pydantic v1 to match FastAPI 0.92
pydantic==1.10.13

# Async HTTP client needed by AWS GPU training utils
aiohttp>=3.8,<4